+++ api/arduino/uno/pin_fn/pin.c.v
@@ -5,27 +5,27 @@
 // Pin management functions Arduino devices
 module pin_fn
 
-@[as_macro] pub const d0 = 0
-@[as_macro] pub const d1 = 1
-@[as_macro] pub const d2 = 2
-@[as_macro] pub const d3 = 3
-@[as_macro] pub const d4 = 4
-@[as_macro] pub const d5 = 5
-@[as_macro] pub const d6 = 6
-@[as_macro] pub const d7 = 7
-@[as_macro] pub const d8 = 8
-@[as_macro] pub const d9 = 9
-@[as_macro] pub const d10 = 10
-@[as_macro] pub const d11 = 11
-@[as_macro] pub const d12 = 12
-@[as_macro] pub const d13 = 13
-@[as_macro] pub const d14 = 14
-@[as_macro] pub const d15 = 15
-@[as_macro] pub const d16 = 16
-@[as_macro] pub const d17 = 17
-@[as_macro] pub const d18 = 18
-@[as_macro] pub const d19 = 19
+@[as_macro] pub const d0 = u8(0)
+@[as_macro] pub const d1 = u8(1)
+@[as_macro] pub const d2 = u8(2)
+@[as_macro] pub const d3 = u8(3)
+@[as_macro] pub const d4 = u8(4)
+@[as_macro] pub const d5 = u8(5)
+@[as_macro] pub const d6 = u8(6)
+@[as_macro] pub const d7 = u8(7)
+@[as_macro] pub const d8 = u8(8)
+@[as_macro] pub const d9 = u8(9)
+@[as_macro] pub const d10 = u8(10)
+@[as_macro] pub const d11 = u8(11)
+@[as_macro] pub const d12 = u8(12)
+@[as_macro] pub const d13 = u8(13)
+@[as_macro] pub const d14 = u8(14)
+@[as_macro] pub const d15 = u8(15)
+@[as_macro] pub const d16 = u8(16)
+@[as_macro] pub const d17 = u8(17)
+@[as_macro] pub const d18 = u8(18)
+@[as_macro] pub const d19 = u8(19)
 
 
 // builtin LED
-@[as_macro]	pub const led0 = 13
+@[as_macro]	pub const led0 = u8(13)
